+++ tests/kebab-case-props-parent.vue
@@ -1,0 +1,10 @@
+<script lang="ts" setup>
+import KebabCasePropsChild from './kebab-case-props-child.vue';
+</script>
+<template>
+    <KebabCasePropsChild
+        NameOne="vue-class-setup"
+        :name-and-age="100"
+        :s-s-s-s="50"
+    />
+</template>